+++ src/board.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdbool.h>
 #include "board.h"
 
 void mostrarTabuleiro(const Tabuleiro *tabAtual)
@@ -110,4 +111,19 @@ void copiar_tabuleiro_para(Tabuleiro *orig, Tabuleiro *dest)
     for (int i = 0; i < orig->linhas; i++)
         for (int j = 0; j < orig->colunas; j++)
             dest->grelha[i][j] = orig->grelha[i][j];
+}
+
+bool compara_tabuleiros(Tabuleiro *tabA, Tabuleiro *tabB)
+{
+    for (int i = 0; i < tabA->linhas; i++)
+    {
+        for (int j = 0; j < tabA->colunas; j++)
+        {
+            if (tabA->grelha[i][j] != tabB->grelha[i][j])
+            {
+                return false;
+            }
+        }
+    }
+    return true;
 }
